feat: create integrations rest client
