+++ chats/apps/api/v1/internal/rest_clients/flows_rest_client.py
@@ -8,7 +8,6 @@ from rest_framework import status
 from chats.apps.api.v1.internal.rest_clients.internal_authorization import (
     InternalAuthentication,
 )
-from chats.core.views import search_dict_list
 
 LOGGER = logging.getLogger(__name__)
 
@@ -151,32 +150,10 @@ class FlowsContactsAndGroupsMixin:
         return groups
 
 
-class FlowsChatGptData:
-    def chatgpt_headers(self, token: str) -> dict:
-        headers = {
-            "Content-Type": "application/json; charset: utf-8",
-            "Authorization": token,
-        }
-        return headers
-
-    def get_chatgpt_token(self, project_uuid: str, token: str):
-        response = requests.get(
-            url=f"{self.base_url}/api/v1/my-apps/?configured=true&project_uuid={project_uuid}",
-            headers=self.chatgpt_headers(token),
-        )
-        try:
-            app_list = response.json()
-            chat_gpt_app = search_dict_list(app_list, "code", "chatgpt")
-            return chat_gpt_app.get("config").get("api_key")
-        except AttributeError:
-            return None
-
-
 class FlowRESTClient(
     InternalAuthentication,
     FlowsContactsAndGroupsMixin,
     FlowsQueueMixin,
-    FlowsChatGptData,
 ):
     def __init__(self, *args, **kwargs):
         self.base_url = settings.FLOWS_API_URL
